test: address pr-review cleanup for MSIX/packaging tests
Test-only refactors from pr-review feedback; product code untouched.

1. PackageLayoutServiceTests: rewrite the FindWinmds dedup test to build a
   genuine duplicate-discovery scenario (same winmd reachable via both the
   recursive References search and the lib search) so it actually guards the
   HashSet dedup instead of just asserting a single unique input.
2. CodeIntegrityCatalogServiceTests: wrap CreateCatalogDefinitionFile .cdf
   outputs in try/finally File.Delete so tests stop leaking temp files.
3. Extract one shared FakeBuildToolsService (Invocations + OnRun + Handler,
   plus an opt-in EmulateSdkToolOutput helper) and reuse it across the
   Bundle/Pri/Runtime/Identity/Orchestration tests, replacing three
   hand-rolled IBuildToolsService fakes.
4. Fold ControllableNugetService's knobs (dependencies, installed-dir,
   return map) into the existing FakeNugetService and use it in
   PackageInstallationServiceTests instead of a parallel fake.

Debug per-file coverage held (11/12 scope files >=95%; CodeIntegrity 93.1%
native blocker unchanged). Full non-E2E suite: 1803 passed / 0 failed /
1 skipped. Release solution build: 0 warnings / 0 errors.
